+++ CHANGELOG.md
@@ -13,13 +13,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities
 
-## [1.4.0] - Upcoming 
+## [1.4.0] - 2025.11.06 
 
 ### Changed
 
 - Updated model meta to not dot arrays
 - Updated permissions to read `TenantPermissions`, `TenantRolePermissions` and to list permissions of `TenantUsers`
 - Updated `Tenants` controller to ensure only the current tenant owner can update the owner if not an admin
+- Updated dependencies
+- Updated documentation
 
 ### Fixed